Refactor Storage
