+++ src/main/java/ok/dht/test/kazakov/dao/Storage.java
@@ -23,16 +23,7 @@ import java.util.concurrent.ThreadFactory;
 
 class Storage implements Closeable {
 
-    private static final Cleaner CLEANER = Cleaner.create(new CleanerThreadFactory());
-
-    private static final long VERSION = 0;
-    private static final int INDEX_HEADER_SIZE = Long.BYTES * 3;
-    private static final int INDEX_RECORD_SIZE = Long.BYTES;
-
-    private static final String FILE_NAME = "data";
-    private static final String FILE_EXT = ".dat";
-    private static final String FILE_EXT_TMP = ".tmp";
-    private static final String COMPACTED_FILE = FILE_NAME + "_compacted_" + FILE_EXT;
+    private static final Cleaner CLEANER = Cleaner.create(new Storage.CleanerThreadFactory());
 
     private static final Logger LOG = LoggerFactory.getLogger(Storage.class);
 
@@ -44,9 +35,9 @@ class Storage implements Closeable {
 
     static Storage load(final Config config) throws IOException {
         final Path basePath = config.basePath();
-        final Path compactedFile = config.basePath().resolve(COMPACTED_FILE);
+        final Path compactedFile = config.basePath().resolve(StorageHelper.COMPACTED_FILE);
         if (Files.exists(compactedFile)) {
-            finishCompact(config, compactedFile);
+            StorageHelper.finishCompact(config, compactedFile);
         }
 
         final ArrayList<MemorySegment> sstables = new ArrayList<>();
@@ -55,9 +46,9 @@ class Storage implements Closeable {
         int i = 0;
         boolean loadedAll = false;
         while (!loadedAll) {
-            final Path nextFile = basePath.resolve(FILE_NAME + i + FILE_EXT);
+            final Path nextFile = basePath.resolve(StorageHelper.FILE_NAME + i + StorageHelper.FILE_EXT);
             try {
-                sstables.add(mapForRead(scope, nextFile));
+                sstables.add(StorageHelper.mapForRead(scope, nextFile));
             } catch (final NoSuchFileException e) {
                 loadedAll = true;
             }
@@ -74,109 +65,9 @@ class Storage implements Closeable {
             final Storage previousState,
             final Collection<Entry<MemorySegment>> entries) throws IOException {
         final int nextSSTableIndex = previousState.sstables.size();
-        final Path sstablePath = config.basePath().resolve(FILE_NAME + nextSSTableIndex + FILE_EXT);
-        save(entries::iterator, sstablePath);
-    }
-
-    private static void save(
-            final Data entries,
-            final Path sstablePath
-    ) throws IOException {
-
-        final Path sstableTmpPath = sstablePath.resolveSibling(sstablePath.getFileName().toString() + FILE_EXT_TMP);
-
-        Files.deleteIfExists(sstableTmpPath);
-        Files.createFile(sstableTmpPath);
-
-        try (@SuppressWarnings("LocalCanBeFinal") ResourceScope writeScope = ResourceScope.newConfinedScope()) {
-            long size = 0;
-            long entriesCount = 0;
-            boolean hasTombstone = false;
-            for (final Entry<MemorySegment> entry : new IterableIteratorWrapper<>(entries.getIterator())) {
-                size += getSize(entry);
-                if (entry.isTombstone()) {
-                    hasTombstone = true;
-                }
-                entriesCount++;
-            }
-
-            final long dataStart = INDEX_HEADER_SIZE + INDEX_RECORD_SIZE * entriesCount;
-
-            final MemorySegment nextSSTable = MemorySegment.mapFile(
-                            sstableTmpPath,
-                            0,
-                            dataStart + size,
-                            FileChannel.MapMode.READ_WRITE,
-                            writeScope
-            );
-
-            long index = 0;
-            long offset = dataStart;
-            for (final Entry<MemorySegment> entry : new IterableIteratorWrapper<>(entries.getIterator())) {
-                MemoryAccess.setLongAtOffset(nextSSTable, INDEX_HEADER_SIZE + index * INDEX_RECORD_SIZE, offset);
-
-                offset += writeRecord(nextSSTable, offset, entry.key());
-                offset += writeRecord(nextSSTable, offset, entry.value());
-
-                index++;
-            }
-
-            MemoryAccess.setLongAtOffset(nextSSTable, 0, VERSION);
-            MemoryAccess.setLongAtOffset(nextSSTable, 8, entriesCount);
-            MemoryAccess.setLongAtOffset(nextSSTable, 16, hasTombstone ? 1 : 0);
-
-            nextSSTable.force();
-        }
-
-        Files.move(sstableTmpPath, sstablePath, StandardCopyOption.ATOMIC_MOVE);
-    }
-
-    private static long getSize(final Entry<MemorySegment> entry) {
-        if (entry.value() == null) {
-            return Long.BYTES + entry.key().byteSize() + Long.BYTES;
-        } else {
-            return Long.BYTES + entry.value().byteSize() + entry.key().byteSize() + Long.BYTES;
-        }
-    }
-
-    public static long getSizeOnDisk(final Entry<MemorySegment> entry) {
-        return getSize(entry) + INDEX_RECORD_SIZE;
-    }
-
-    private static long writeRecord(final MemorySegment nextSSTable, final long offset, final MemorySegment record) {
-        if (record == null) {
-            MemoryAccess.setLongAtOffset(nextSSTable, offset, -1);
-            return Long.BYTES;
-        }
-        final long recordSize = record.byteSize();
-        MemoryAccess.setLongAtOffset(nextSSTable, offset, recordSize);
-        nextSSTable.asSlice(offset + Long.BYTES, recordSize).copyFrom(record);
-        return Long.BYTES + recordSize;
-    }
-
-    @SuppressWarnings("DuplicateThrows")
-    private static MemorySegment mapForRead(final ResourceScope scope,
-                                            final Path file) throws NoSuchFileException, IOException {
-        final long size = Files.size(file);
-
-        return MemorySegment.mapFile(file, 0, size, FileChannel.MapMode.READ_ONLY, scope);
-    }
-
-    public static void compact(final Config config, final Data data) throws IOException {
-        final Path compactedFile = config.basePath().resolve(COMPACTED_FILE);
-        save(data, compactedFile);
-        finishCompact(config, compactedFile);
-    }
-
-    private static void finishCompact(final Config config, final Path compactedFile) throws IOException {
-        for (int i = 0; ; i++) {
-            final Path nextFile = config.basePath().resolve(FILE_NAME + i + FILE_EXT);
-            if (!Files.deleteIfExists(nextFile)) {
-                break;
-            }
-        }
-
-        Files.move(compactedFile, config.basePath().resolve(FILE_NAME + 0 + FILE_EXT), StandardCopyOption.ATOMIC_MOVE);
+        final Path sstablePath = config.basePath()
+                .resolve(StorageHelper.FILE_NAME + nextSSTableIndex + StorageHelper.FILE_EXT);
+        StorageHelper.save(entries::iterator, sstablePath);
     }
 
     private Storage(final ResourceScope scope, final List<MemorySegment> sstables, final boolean hasTombstones) {
@@ -211,7 +102,8 @@ class Storage implements Closeable {
         while (left <= right) {
             final long mid = (left + right) >>> 1;
 
-            final long keyPos = MemoryAccess.getLongAtOffset(sstable, INDEX_HEADER_SIZE + mid * INDEX_RECORD_SIZE);
+            final long keyPos = MemoryAccess.getLongAtOffset(sstable,
+                    StorageHelper.INDEX_HEADER_SIZE + mid * StorageHelper.INDEX_RECORD_SIZE);
             final long keySize = MemoryAccess.getLongAtOffset(sstable, keyPos);
 
             final MemorySegment keyForCheck = sstable.asSlice(keyPos + Long.BYTES, keySize);
@@ -230,7 +122,8 @@ class Storage implements Closeable {
 
     private Entry<MemorySegment> entryAt(final MemorySegment sstable, final long keyIndex) {
         try {
-            final long offset = MemoryAccess.getLongAtOffset(sstable, INDEX_HEADER_SIZE + keyIndex * INDEX_RECORD_SIZE);
+            final long offset = MemoryAccess.getLongAtOffset(sstable,
+                    StorageHelper.INDEX_HEADER_SIZE + keyIndex * StorageHelper.INDEX_RECORD_SIZE);
             final long keySize = MemoryAccess.getLongAtOffset(sstable, offset);
             final long valueOffset = offset + Long.BYTES + keySize;
             final long valueSize = MemoryAccess.getLongAtOffset(sstable, valueOffset);
