+++ src/main/java/ok/dht/test/kazakov/dao/StorageHelper.java
@@ -1,0 +1,131 @@
+package ok.dht.test.kazakov.dao;
+
+import jdk.incubator.foreign.MemoryAccess;
+import jdk.incubator.foreign.MemorySegment;
+import jdk.incubator.foreign.ResourceScope;
+
+import javax.annotation.Nonnull;
+import java.io.IOException;
+import java.lang.ref.Cleaner;
+import java.nio.channels.FileChannel;
+import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.concurrent.ThreadFactory;
+
+public class StorageHelper {
+
+    private static final long VERSION = 0;
+
+    static final String FILE_NAME = "data";
+    static final String FILE_EXT = ".dat";
+    private static final String FILE_EXT_TMP = ".tmp";
+    static final String COMPACTED_FILE = FILE_NAME + "_compacted_" + FILE_EXT;
+
+    static final int INDEX_HEADER_SIZE = Long.BYTES * 3;
+    static final int INDEX_RECORD_SIZE = Long.BYTES;
+
+    static void save(
+            final Storage.Data entries,
+            final Path sstablePath
+    ) throws IOException {
+
+        final Path sstableTmpPath = sstablePath.resolveSibling(sstablePath.getFileName().toString() + FILE_EXT_TMP);
+
+        Files.deleteIfExists(sstableTmpPath);
+        Files.createFile(sstableTmpPath);
+
+        try (@SuppressWarnings("LocalCanBeFinal") ResourceScope writeScope = ResourceScope.newConfinedScope()) {
+            long size = 0;
+            long entriesCount = 0;
+            boolean hasTombstone = false;
+            for (final Entry<MemorySegment> entry : new IterableIteratorWrapper<>(entries.getIterator())) {
+                size += getSize(entry);
+                if (entry.isTombstone()) {
+                    hasTombstone = true;
+                }
+                entriesCount++;
+            }
+
+            final long dataStart = INDEX_HEADER_SIZE + INDEX_RECORD_SIZE * entriesCount;
+
+            final MemorySegment nextSSTable = MemorySegment.mapFile(
+                    sstableTmpPath,
+                    0,
+                    dataStart + size,
+                    FileChannel.MapMode.READ_WRITE,
+                    writeScope
+            );
+
+            long index = 0;
+            long offset = dataStart;
+            for (final Entry<MemorySegment> entry : new IterableIteratorWrapper<>(entries.getIterator())) {
+                MemoryAccess.setLongAtOffset(nextSSTable, INDEX_HEADER_SIZE + index * INDEX_RECORD_SIZE, offset);
+
+                offset += writeRecord(nextSSTable, offset, entry.key());
+                offset += writeRecord(nextSSTable, offset, entry.value());
+
+                index++;
+            }
+
+            MemoryAccess.setLongAtOffset(nextSSTable, 0, VERSION);
+            MemoryAccess.setLongAtOffset(nextSSTable, 8, entriesCount);
+            MemoryAccess.setLongAtOffset(nextSSTable, 16, hasTombstone ? 1 : 0);
+
+            nextSSTable.force();
+        }
+
+        Files.move(sstableTmpPath, sstablePath, StandardCopyOption.ATOMIC_MOVE);
+    }
+
+    private static long getSize(final Entry<MemorySegment> entry) {
+        if (entry.value() == null) {
+            return Long.BYTES + entry.key().byteSize() + Long.BYTES;
+        } else {
+            return Long.BYTES + entry.value().byteSize() + entry.key().byteSize() + Long.BYTES;
+        }
+    }
+
+    public static long getSizeOnDisk(final Entry<MemorySegment> entry) {
+        return getSize(entry) + INDEX_RECORD_SIZE;
+    }
+
+    private static long writeRecord(final MemorySegment nextSSTable, final long offset, final MemorySegment record) {
+        if (record == null) {
+            MemoryAccess.setLongAtOffset(nextSSTable, offset, -1);
+            return Long.BYTES;
+        }
+        final long recordSize = record.byteSize();
+        MemoryAccess.setLongAtOffset(nextSSTable, offset, recordSize);
+        nextSSTable.asSlice(offset + Long.BYTES, recordSize).copyFrom(record);
+        return Long.BYTES + recordSize;
+    }
+
+    @SuppressWarnings("DuplicateThrows")
+    static MemorySegment mapForRead(final ResourceScope scope,
+                                            final Path file) throws NoSuchFileException, IOException {
+        final long size = Files.size(file);
+
+        return MemorySegment.mapFile(file, 0, size, FileChannel.MapMode.READ_ONLY, scope);
+    }
+
+    public static void compact(final Config config, final Storage.Data data) throws IOException {
+        final Path compactedFile = config.basePath().resolve(COMPACTED_FILE);
+        save(data, compactedFile);
+        finishCompact(config, compactedFile);
+    }
+
+    static void finishCompact(final Config config, final Path compactedFile) throws IOException {
+        for (int i = 0; ; i++) {
+            final Path nextFile = config.basePath().resolve(FILE_NAME + i + FILE_EXT);
+            if (!Files.deleteIfExists(nextFile)) {
+                break;
+            }
+        }
+
+        Files.move(compactedFile, config.basePath().resolve(FILE_NAME + 0 + FILE_EXT), StandardCopyOption.ATOMIC_MOVE);
+    }
+}
